requirements.txt
Django==4.2.4
Pillow==10.0.0
numpy==1.25.2
django-allauth==0.54.0
python-decouple==3.8
django-environ==0.10.0
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dj-rest-auth==5.0.1
